[Reporting ] run task fn simplifications
