+++ packages/kbn-reporting/export_types/pdf/printable_pdf_v2.ts
@@ -22,17 +22,18 @@ import {
   REPORTING_REDIRECT_LOCATOR_STORE_KEY,
   REPORTING_TRANSACTION_TYPE,
 } from '@kbn/reporting-common';
-import type { TaskRunResult, UrlOrUrlLocatorTuple } from '@kbn/reporting-common/types';
+import type { TaskRunResult } from '@kbn/reporting-common/types';
+import type { TaskPayloadPDFV2 } from '@kbn/reporting-export-types-pdf-common';
 import {
   JobParamsPDFV2,
   PDF_JOB_TYPE_V2,
   PDF_REPORT_TYPE_V2,
-  TaskPayloadPDFV2,
 } from '@kbn/reporting-export-types-pdf-common';
-import { decryptJobHeaders, getFullRedirectAppUrl, ExportType } from '@kbn/reporting-server';
+import { ExportType, decryptJobHeaders, getFullRedirectAppUrl } from '@kbn/reporting-server';
+import type { UrlOrUrlWithContext } from '@kbn/screenshotting-plugin/server/screenshots';
 
-import { generatePdfObservableV2 } from './generate_pdf_v2';
 import { getCustomLogo } from './get_custom_logo';
+import { getTracker } from './pdf_tracker';
 
 export class PdfExportType extends ExportType<JobParamsPDFV2, TaskPayloadPDFV2> {
   id = PDF_REPORT_TYPE_V2;
@@ -80,66 +81,82 @@ export class PdfExportType extends ExportType<JobParamsPDFV2, TaskPayloadPDFV2> 
     cancellationToken: CancellationToken,
     stream: Writable
   ) => {
-    const jobLogger = this.logger.get(`execute-job:${jobId}`);
+    const logger = this.logger.get(`execute-job:${jobId}`);
     const apmTrans = apm.startTransaction('execute-job-pdf-v2', REPORTING_TRANSACTION_TYPE);
     const apmGetAssets = apmTrans.startSpan('get-assets', 'setup');
     let apmGeneratePdf: { end: () => void } | null | undefined;
     const { encryptionKey } = this.config;
 
     const process$: Rx.Observable<TaskRunResult> = of(1).pipe(
-      mergeMap(() => decryptJobHeaders(encryptionKey, payload.headers, jobLogger)),
+      mergeMap(() => decryptJobHeaders(encryptionKey, payload.headers, logger)),
       mergeMap(async (headers: Headers) => {
-        const fakeRequest = this.getFakeRequest(headers, payload.spaceId, jobLogger);
+        const fakeRequest = this.getFakeRequest(headers, payload.spaceId, logger);
         const uiSettingsClient = await this.getUiSettingsClient(fakeRequest);
         return await getCustomLogo(uiSettingsClient, headers);
       }),
       mergeMap(({ logo, headers }) => {
         const { browserTimezone, layout, title, locatorParams } = payload;
-        let urls: UrlOrUrlLocatorTuple[];
-        if (locatorParams) {
-          urls = locatorParams.map((locator) => [
-            getFullRedirectAppUrl(
-              this.config,
-              this.getServerInfo(),
-              payload.spaceId,
-              payload.forceNow
-            ),
-            locator,
-          ]);
-        }
 
         apmGetAssets?.end();
 
         apmGeneratePdf = apmTrans.startSpan('generate-pdf-pipeline', 'execute');
 
-        return generatePdfObservableV2(
-          this.config,
-          this.getServerInfo(),
-          () =>
-            this.startDeps.screenshotting!.getScreenshots({
-              format: 'pdf',
-              title,
-              logo,
-              browserTimezone,
-              headers,
-              layout,
-              urls: urls.map((url) =>
-                typeof url === 'string'
-                  ? url
-                  : [url[0], { [REPORTING_REDIRECT_LOCATOR_STORE_KEY]: url[1] }]
-              ),
-            }),
-          payload,
-          locatorParams,
-          {
+        const tracker = getTracker();
+        tracker.startScreenshots();
+
+        /**
+         * For each locator we get the relative URL to the redirect app
+         */
+        const urls = locatorParams.map((locator) => [
+          getFullRedirectAppUrl(
+            this.config,
+            this.getServerInfo(),
+            payload.spaceId,
+            payload.forceNow
+          ),
+          locator,
+        ]) as unknown as UrlOrUrlWithContext[];
+
+        return this.startDeps
+          .screenshotting!.getScreenshots({
             format: 'pdf',
             title,
             logo,
             browserTimezone,
             headers,
             layout,
-          }
-        );
+            urls: urls.map((url) =>
+              typeof url === 'string'
+                ? url
+                : [url[0], { [REPORTING_REDIRECT_LOCATOR_STORE_KEY]: url[1] }]
+            ),
+          })
+          .pipe(
+            tap(({ metrics }) => {
+              if (metrics.cpu) {
+                tracker.setCpuUsage(metrics.cpu);
+              }
+              if (metrics.memory) {
+                tracker.setMemoryUsage(metrics.memory);
+              }
+            }),
+            mergeMap(async ({ data: buffer, errors, metrics, renderErrors }) => {
+              tracker.endScreenshots();
+              const warnings: string[] = [];
+              if (errors) {
+                warnings.push(...errors.map((error) => error.message));
+              }
+              if (renderErrors) {
+                warnings.push(...renderErrors);
+              }
+
+              return {
+                buffer,
+                metrics,
+                warnings,
+              };
+            })
+          );
       }),
       tap(({ buffer }) => {
         apmGeneratePdf?.end();
@@ -154,7 +171,7 @@ export class PdfExportType extends ExportType<JobParamsPDFV2, TaskPayloadPDFV2> 
         warnings,
       })),
       catchError((err) => {
-        jobLogger.error(err);
+        logger.error(err);
         return Rx.throwError(() => err);
       })
     );
