+++ packages/kbn-reporting/export_types/png/png_v2.ts
@@ -38,9 +38,7 @@ import {
   PNG_REPORT_TYPE_V2,
   TaskPayloadPNGV2,
 } from '@kbn/reporting-export-types-png-common';
-import { decryptJobHeaders, getFullRedirectAppUrl, ExportType } from '@kbn/reporting-server';
-
-import { generatePngObservable } from './generate_png';
+import { decryptJobHeaders, ExportType, getFullRedirectAppUrl } from '@kbn/reporting-server';
 
 export class PngExportType extends ExportType<JobParamsPNGV2, TaskPayloadPNGV2> {
   id = PNG_REPORT_TYPE_V2;
@@ -88,14 +86,14 @@ export class PngExportType extends ExportType<JobParamsPNGV2, TaskPayloadPNGV2> 
     cancellationToken: CancellationToken,
     stream: Writable
   ) => {
-    const jobLogger = this.logger.get(`execute-job:${jobId}`);
+    const logger = this.logger.get(`execute-job:${jobId}`);
     const apmTrans = apm.startTransaction('execute-job-pdf-v2', REPORTING_TRANSACTION_TYPE);
     const apmGetAssets = apmTrans.startSpan('get-assets', 'setup');
     let apmGeneratePng: { end: () => void } | null | undefined;
     const { encryptionKey } = this.config;
 
     const process$: Observable<TaskRunResult> = of(1).pipe(
-      mergeMap(() => decryptJobHeaders(encryptionKey, payload.headers, jobLogger)),
+      mergeMap(() => decryptJobHeaders(encryptionKey, payload.headers, logger)),
       mergeMap((headers) => {
         const url = getFullRedirectAppUrl(
           this.config,
@@ -108,22 +106,57 @@ export class PngExportType extends ExportType<JobParamsPNGV2, TaskPayloadPNGV2> 
 
         apmGetAssets?.end();
         apmGeneratePng = apmTrans.startSpan('generate-png-pipeline', 'execute');
+        const options = {
+          headers,
+          browserTimezone: payload.browserTimezone,
+          layout: { ...payload.layout, id: 'preserve_layout' as const },
+        };
 
-        return generatePngObservable(
-          () =>
-            this.startDeps.screenshotting!.getScreenshots({
-              format: 'png',
-              headers,
-              layout: { ...payload.layout, id: 'preserve_layout' },
-              urls: [[url, { [REPORTING_REDIRECT_LOCATOR_STORE_KEY]: locatorParams }]],
-            }),
-          jobLogger,
-          {
+        if (!options.layout?.dimensions) {
+          throw new Error(`LayoutParams.Dimensions is undefined.`);
+        }
+
+        const apmScreenshots = apmTrans?.startSpan('screenshots-pipeline', 'setup');
+        let apmBuffer: typeof apm.currentSpan;
+
+        return this.startDeps
+          .screenshotting!.getScreenshots({
+            format: 'png',
             headers,
-            browserTimezone: payload.browserTimezone,
             layout: { ...payload.layout, id: 'preserve_layout' },
-          }
-        );
+            urls: [[url, { [REPORTING_REDIRECT_LOCATOR_STORE_KEY]: locatorParams }]],
+          })
+          .pipe(
+            tap(({ metrics }) => {
+              if (metrics) {
+                apmTrans.setLabel('cpu', metrics.cpu, false);
+                apmTrans.setLabel('memory', metrics.memory, false);
+              }
+              apmScreenshots?.end();
+              apmBuffer = apmTrans.startSpan('get-buffer', 'output') ?? null;
+            }),
+            map(({ metrics, results }) => ({
+              metrics,
+              buffer: results[0].screenshots[0].data,
+              warnings: results.reduce((found, current) => {
+                if (current.error) {
+                  found.push(current.error.message);
+                }
+                if (current.renderErrors) {
+                  found.push(...current.renderErrors);
+                }
+                return found;
+              }, [] as string[]),
+            })),
+            tap(({ buffer }) => {
+              logger.debug(`PNG buffer byte length: ${buffer.byteLength}`);
+              apmTrans.setLabel('byte-length', buffer.byteLength, false);
+            }),
+            finalize(() => {
+              apmBuffer?.end();
+              apmTrans.end();
+            })
+          );
       }),
       tap(({ buffer }) => stream.write(buffer)),
       map(({ metrics, warnings }) => ({
@@ -131,7 +164,7 @@ export class PngExportType extends ExportType<JobParamsPNGV2, TaskPayloadPNGV2> 
         metrics: { png: metrics },
         warnings,
       })),
-      tap({ error: (error) => jobLogger.error(error) }),
+      tap({ error: (error) => logger.error(error) }),
       finalize(() => apmGeneratePng?.end())
     );
